+++ src/virtual-keyboard/variants.ts
@@ -9,6 +9,7 @@ import { VirtualKeyboard } from './virtual-keyboard';
 import { FOREGROUND_COLORS, BACKGROUND_COLORS } from '../core/color';
 import MathfieldElement from '../public/mathfield-element';
 import type { VirtualKeyboardKeycap } from '../public/virtual-keyboard';
+import { getComputedDir } from 'ui/i18n/utils';
 
 const VARIANTS: {
   [variantID: string]: (string | Partial<VirtualKeyboardKeycap>)[];
@@ -237,7 +238,11 @@ export function showVariantsPanel(
       )
     );
     const top = position.top - variantPanel.clientHeight + 5;
-    variantPanel.style.transform = `translate(${left}px, ${top}px)`;
+
+    console.log('left: ', left);
+
+    variantPanel.style.left = `${left}px`;
+    variantPanel.style.top = `${top}px`;
     variantPanel.classList.add('is-visible');
     // Add the events handlers (which may dismiss the panel) only after the
     // panel has been displayed